docs: Add JS docs to createRecurringEvents.js
